refactor(server): split sio-state.ts into 11 focused modules
Extract the 1233-line Redis CRUD monolith into focused modules under
packages/server/src/ws/sio-state/:

- types.ts        — interfaces (RedisSessionData, etc.) + TTL constants
- client.ts       — Redis client lifecycle (init/get/close/require)
- keys.ts         — all 15 key builder functions + KEY_PREFIX
- serialization.ts — toHashFields, parse* functions, SESSION_SUMMARY_FIELDS
- sessions.ts     — Session CRUD + incrementSeq/getSeq
- runners.ts      — Runner CRUD + runner association functions
- terminals.ts    — Terminal CRUD
- child-sessions.ts — Child tracking, delink markers, pending delink functions
- push.ts         — Push pending question state with Lua scripts
- cleanup.ts      — scanExpiredSessions, cleanStaleIndexEntries
- index.ts        — Barrel re-exports of all public APIs

All existing imports updated from sio-state.js → sio-state/index.js.
Also fixes two pre-existing implicit-any errors in cleanStaleIndexEntries.
