+++ packages/server/src/ws/sio-registry/meta.ts
@@ -6,7 +6,7 @@
 // ============================================================================
 
 import { defaultMetaState, type SessionMetaState, type MetaRelayEvent } from "@pizzapi/protocol";
-import { getSession, updateSessionFields } from "../sio-state.js";
+import { getSession, updateSessionFields } from "../sio-state/index.js";
 import { getIo } from "./context.js";
 import { createLogger } from "@pizzapi/tools";
