+++ packages/server/src/ws/sio-state/index.ts
@@ -1,0 +1,12 @@
+// ============================================================================
+// sio-state/index.ts — Barrel re-exports of all public APIs
+// ============================================================================
+
+export * from "./types.js";
+export * from "./client.js";
+export * from "./sessions.js";
+export * from "./runners.js";
+export * from "./terminals.js";
+export * from "./child-sessions.js";
+export * from "./push.js";
+export * from "./cleanup.js";
